+++ src/metadata/modules_sources.h
@@ -77,7 +77,7 @@ struct method_data_t_hash
 {
     size_t operator()(const method_data_t &p) const
     {
-        return p.methodDef + (uint32_t)p.startLine * 100 + (uint32_t)p.endLine * 1000;
+        return (size_t)p.methodDef + (size_t)p.startLine * 100 + (size_t)p.endLine * 1000;
     }
 };
